requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.15.12
botocore==1.18.12
colorama==0.4.3
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-storages==1.10.1
gunicorn==20.0.4
isort==5.5.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
